requirements.txt
streamlit>=1.25
ultralytics>=8.0.0
torch>=2.0.0
opencv-python-headless
numpy
Pillow
requests
